refactor: simplify address review findings on Profiler tool surface
Apply 11 of 12 review-pass findings on PR #773. Major changes: SaveData composes JSON via sibling Get* methods so the snapshot stays in sync with each POCO and the missing frame block + per-section fields are now populated; LoadData caps input at 10 MB to bound worst-case memory; Profiler.cs harmonises EnabledModules/AvailableModules visibility (both internal) and exposes AvailableModules as IReadOnlyList; GetStatus/GetMemoryStats/GetScriptStats divide long counters in double precision to preserve precision above 16 MB; Start now warns when ExecuteMenuItem returns false instead of silently no-opping; CaptureFrame drops the unused frameCount parameter for the established no-arg sentinel; tests snapshot+restore the static EnabledModules set across SetUp/TearDown. Deferred: the nit renaming string nothing=null to _unused since that placeholder matches Tool_Console.ClearLogs and the surrounding convention.

simplify-pass: 1

+++ Unity-MCP-Plugin/Packages/com.ivanmurzak.unity.mcp/Editor/Scripts/API/Tool/Profiler.GetMemoryStats.cs
@@ -44,17 +44,19 @@ namespace com.IvanMurzak.Unity.MCP.Editor.API
         [Description("Returns memory statistics from the Unity Profiler (all values in MB).")]
         public MemoryStatsData GetMemoryStats(string? nothing = null)
         {
+            // Divide in double precision then cast — `long / 1048576f` loses precision
+            // above ~16 MB because the float mantissa is only 24 bits.
             return MainThread.Instance.Run(() => new MemoryStatsData
             {
-                TotalReservedMemoryMB = UnityProfiler.GetTotalReservedMemoryLong() / 1048576f,
-                TotalAllocatedMemoryMB = UnityProfiler.GetTotalAllocatedMemoryLong() / 1048576f,
-                TotalUnusedReservedMemoryMB = UnityProfiler.GetTotalUnusedReservedMemoryLong() / 1048576f,
-                MonoHeapSizeMB = UnityProfiler.GetMonoHeapSizeLong() / 1048576f,
-                MonoUsedSizeMB = UnityProfiler.GetMonoUsedSizeLong() / 1048576f,
-                TempAllocatorSizeMB = UnityProfiler.GetTempAllocatorSize() / 1048576f,
-                GraphicsMemoryMB = UnityProfiler.GetAllocatedMemoryForGraphicsDriver() / 1048576f,
-                MaxUsedMemoryMB = UnityProfiler.maxUsedMemory / 1048576f,
-                UsedHeapSizeMB = UnityProfiler.usedHeapSizeLong / 1048576f
+                TotalReservedMemoryMB = (float)(UnityProfiler.GetTotalReservedMemoryLong() / 1048576.0),
+                TotalAllocatedMemoryMB = (float)(UnityProfiler.GetTotalAllocatedMemoryLong() / 1048576.0),
+                TotalUnusedReservedMemoryMB = (float)(UnityProfiler.GetTotalUnusedReservedMemoryLong() / 1048576.0),
+                MonoHeapSizeMB = (float)(UnityProfiler.GetMonoHeapSizeLong() / 1048576.0),
+                MonoUsedSizeMB = (float)(UnityProfiler.GetMonoUsedSizeLong() / 1048576.0),
+                TempAllocatorSizeMB = (float)(UnityProfiler.GetTempAllocatorSize() / 1048576.0),
+                GraphicsMemoryMB = (float)(UnityProfiler.GetAllocatedMemoryForGraphicsDriver() / 1048576.0),
+                MaxUsedMemoryMB = (float)(UnityProfiler.maxUsedMemory / 1048576.0),
+                UsedHeapSizeMB = (float)(UnityProfiler.usedHeapSizeLong / 1048576.0)
             });
         }
     }
